Added endorsements_count

+++ src/main/java/com/project/talent1/Repositories/UsersHasTalentsRepository.java
@@ -10,5 +10,5 @@ public interface UsersHasTalentsRepository extends CrudRepository<Users_has_tale
     Iterable<Users_has_talents> findAll();
 
     @Query(value = "SELECT * FROM users_has_talents WHERE person_id = ?1 AND talent_id = ?2", nativeQuery = true)
-    Users_has_talents findByPersonIdTalentId(int person_id, int talent_id);
+    Users_has_talents findByPersonIdTalentId(long person_id, long talent_id);
 }
